+++ NEWS.md
@@ -1,6 +1,6 @@
 # ggpedigree 0.5.0
 * Added segment_linetype options to the `ggpedigree` function.
-
+* Added usage of color_pallette to the `ggpedigree` function.
 # ggpedigree 0.4.1
 ## Status: Submitted to CRAN
 * Fixed a bug in the `ggpedigree` function that caused an error when using a custom ID name and requesting the plot to be returned as a ggplot object.